Development merge with main

+++ API-SOS_Code/Controllers/IngredientController.cs
@@ -43,31 +43,40 @@ namespace API_SOS_Code.Controllers
         [HttpPost]
         public async Task<ActionResult<Ingredient>> Add(InsertIngredientDTO ingredientDTO)
         {
-            try
-            {
-                if (ingredientDTO == null) return BadRequest("Ingredient data is required!");
+ 
+            if (ingredientDTO == null) return BadRequest("Ingredient data is required!");
 
-                var ingredient = new Ingredient
+            var ingredients = new List<Ingredient>();
+            for (int i = 0; i < ingredientDTO.Quantity; i++)
+            {
+                ingredients.Add(new Ingredient
                 {
                     Name = ingredientDTO.Name,
                     ExpirationDate = ingredientDTO.ExpirationDate
-                };
-                _context.Ingredients.Add(ingredient);
+                });
+            }
 
-                // Check if the ingredient name already exists
-                var ingredientsName = await _context.IngredientsName.ToListAsync();
+            await _context.Ingredients.AddRangeAsync(ingredients);
+
+            // Check if the ingredient name already exists
+            var ingredientsName = await _context.IngredientsName.ToListAsync();
+            foreach (var ingredient in ingredients)
+            {
                 if (ingredientsName.All(i => i.Name != ingredient.Name))
                 {
                     var newIngredientName = new IngredientsName
                     {
                         Name = ingredient.Name
                     };
-                    _context.IngredientsName.Add(newIngredientName);
+                    ingredientsName.Add(newIngredientName);
+                    await _context.IngredientsName.AddAsync(newIngredientName);
                 }
+            }
 
+            try
+            {
                 await _context.SaveChangesAsync();
-
-                return CreatedAtAction(nameof(GetById), new { id = ingredient.Id } , ingredient);
+                return CreatedAtAction(nameof(GetAll), ingredients);
             }
             catch (DbUpdateException)
             {
@@ -79,10 +88,12 @@ namespace API_SOS_Code.Controllers
         [HttpDelete("{id}")]
         public async Task<ActionResult<Ingredient>> Delete(int id)
         {
+            if (id <= 0) return BadRequest("Invalid ingredient ID!");
+
             try
             {
                 var ingredient = await _context.Ingredients.FirstOrDefaultAsync(i => i.Id == id);
-                
+
                 if (ingredient == null) return NotFound($"Ingredient {id} not found!");
 
                 _context.Ingredients.Remove(ingredient);
@@ -92,6 +103,36 @@ namespace API_SOS_Code.Controllers
             catch (DbUpdateException)
             {
                 return BadRequest($"An error ocurred while deleting the ingredient with id {id}");
+            }
+        }
+
+        [Authorize(Roles = "Admin")]
+        [HttpDelete]
+        public async Task<ActionResult<IEnumerable<Ingredient>>> DeleteBatch([FromBody] IEnumerable<int> ids)
+        {
+            if (ids == null || !ids.Any()) return BadRequest("No ingredient IDs provided for deletion!");
+
+            try
+            {
+                var ingredients = new List<Ingredient>();
+
+                foreach (var ingredientId in ids)
+                {
+                    var ingredient = await _context.Ingredients.FirstOrDefaultAsync(i => i.Id == ingredientId);
+                    if (ingredient != null)
+                    {
+                        ingredients.Add(ingredient);
+                    }
+                }
+
+                ingredients.ForEach(ingredient => _context.Ingredients.Remove(ingredient));
+
+                await _context.SaveChangesAsync();
+                return Ok(ingredients);
+            }
+            catch (DbUpdateException)
+            {
+                return BadRequest($"An error ocurred while deleting the ingredient with id {ids}");
             }
         }
